refactor: 그룹 회원관련 로직을 GroupUsers로 옮긴다
